core-model-prediction/requirements.txt
torch
nltk
vaderSentiment
pandas
textstat
scikit-learn==1.4.1.post1
transformers
fastapi
uvicorn
google-cloud-secret-manager